Extract PropertyField/Value transformers and handlers into properties.ts
PropertyField and PropertyValue are shared primitives, not boards-specific
(future features such as cards and content flagging will reuse them).
Moving them out of the boards files keeps the boards namespace focused on
BoardView and lets other product features import the property handlers
without a boards dependency.

- transformers/properties.ts: transformPropertyFieldRecord, transformPropertyValueRecord
- handlers/properties.ts: PropertiesHandler mixin + PropertiesHandlerMix
- boards.ts files keep only BoardView code
- ServerDataOperator now composes PropertiesHandler alongside BoardsHandler
- Tests split into properties.test.ts and boards.test.ts accordingly
