+++ app/database/operator/server_data_operator/handlers/boards.test.ts
@@ -3,15 +3,11 @@
 
 import {MM_TABLES} from '@constants/database';
 import DatabaseManager from '@database/manager';
-import {
-    transformPropertyFieldRecord,
-    transformPropertyValueRecord,
-    transformBoardViewRecord,
-} from '@database/operator/server_data_operator/transformers/boards';
+import {transformBoardViewRecord} from '@database/operator/server_data_operator/transformers/boards';
 
 import type ServerDataOperator from '@database/operator/server_data_operator';
 
-const {PROPERTY_FIELD, PROPERTY_VALUE, BOARD_VIEW} = MM_TABLES.SERVER;
+const {BOARD_VIEW} = MM_TABLES.SERVER;
 
 describe('*** Operator: Boards Handlers tests ***', () => {
     let operator: ServerDataOperator;
@@ -75,96 +71,6 @@ describe('*** Operator: Boards Handlers tests ***', () => {
 
             const spy = jest.spyOn(operator, 'handleRecords');
             const result = await operator.handleBoardViews({prepareRecordsOnly: false});
-
-            expect(result).toEqual([]);
-            expect(spy).not.toHaveBeenCalled();
-        });
-    });
-
-    describe('=> handlePropertyFields', () => {
-        it('should delegate to handleRecords with the property field transformer', async () => {
-            expect.assertions(2);
-
-            const spy = jest.spyOn(operator, 'handleRecords');
-            const fields: PropertyField[] = [{
-                id: 'field1',
-                group_id: 'group1',
-                name: 'Status',
-                type: 'select',
-                object_type: 'card',
-                target_id: 'channel1',
-                target_type: 'channel',
-                protected: false,
-                create_at: 1,
-                update_at: 1,
-                delete_at: 0,
-                created_by: 'user1',
-                updated_by: 'user1',
-            }];
-
-            await operator.handlePropertyFields({fields, prepareRecordsOnly: false});
-
-            expect(spy).toHaveBeenCalledTimes(1);
-            expect(spy).toHaveBeenCalledWith({
-                fieldName: 'id',
-                createOrUpdateRawValues: fields,
-                tableName: PROPERTY_FIELD,
-                prepareRecordsOnly: false,
-                transformer: transformPropertyFieldRecord,
-            }, 'handlePropertyFields');
-        });
-
-        it('should return [] without calling handleRecords when fields is empty', async () => {
-            expect.assertions(2);
-
-            const spy = jest.spyOn(operator, 'handleRecords');
-            const result = await operator.handlePropertyFields({fields: [], prepareRecordsOnly: false});
-
-            expect(result).toEqual([]);
-            expect(spy).not.toHaveBeenCalled();
-        });
-    });
-
-    describe('=> handlePropertyValues', () => {
-        it('should delegate to handleRecords with the property value transformer and prepareRecordsOnly: true', async () => {
-            expect.assertions(2);
-
-            const originalHandleRecords = operator.handleRecords;
-            operator.handleRecords = jest.fn().mockResolvedValue([]);
-
-            const values: PropertyValue[] = [{
-                id: 'value1',
-                field_id: 'field1',
-                target_id: 'post1',
-                target_type: 'post',
-                group_id: 'group1',
-                value: 'Done',
-                create_at: 1,
-                update_at: 1,
-                delete_at: 0,
-                created_by: 'user1',
-                updated_by: 'user1',
-            }];
-
-            await operator.handlePropertyValues({values, prepareRecordsOnly: true});
-
-            expect(operator.handleRecords).toHaveBeenCalledTimes(1);
-            expect(operator.handleRecords).toHaveBeenCalledWith({
-                fieldName: 'id',
-                createOrUpdateRawValues: values,
-                tableName: PROPERTY_VALUE,
-                prepareRecordsOnly: true,
-                transformer: transformPropertyValueRecord,
-            }, 'handlePropertyValues');
-
-            operator.handleRecords = originalHandleRecords;
-        });
-
-        it('should return [] without calling handleRecords when values is empty', async () => {
-            expect.assertions(2);
-
-            const spy = jest.spyOn(operator, 'handleRecords');
-            const result = await operator.handlePropertyValues({values: [], prepareRecordsOnly: false});
 
             expect(result).toEqual([]);
             expect(spy).not.toHaveBeenCalled();
