+++ app/database/operator/server_data_operator/handlers/boards.ts
@@ -2,28 +2,18 @@
 // See LICENSE.txt for license information.
 
 import {MM_TABLES} from '@constants/database';
-import {
-    transformPropertyFieldRecord,
-    transformPropertyValueRecord,
-    transformBoardViewRecord,
-} from '@database/operator/server_data_operator/transformers/boards';
+import {transformBoardViewRecord} from '@database/operator/server_data_operator/transformers/boards';
 import {getUniqueRawsBy} from '@database/operator/utils/general';
 import {logWarning} from '@utils/log';
 
 import type ServerDataOperatorBase from '.';
 import type Model from '@nozbe/watermelondb/Model';
-import type {
-    HandlePropertyFieldsArgs,
-    HandlePropertyValuesArgs,
-    HandleBoardViewsArgs,
-} from '@typings/database/database';
+import type {HandleBoardViewsArgs} from '@typings/database/database';
 
-const {PROPERTY_FIELD, PROPERTY_VALUE, BOARD_VIEW} = MM_TABLES.SERVER;
+const {BOARD_VIEW} = MM_TABLES.SERVER;
 
 export interface BoardsHandlerMix {
     handleBoardViews: ({boardViews, prepareRecordsOnly}: HandleBoardViewsArgs) => Promise<Model[]>;
-    handlePropertyFields: ({fields, prepareRecordsOnly}: HandlePropertyFieldsArgs) => Promise<Model[]>;
-    handlePropertyValues: ({values, prepareRecordsOnly}: HandlePropertyValuesArgs) => Promise<Model[]>;
 }
 
 const BoardsHandler = <TBase extends Constructor<ServerDataOperatorBase>>(superclass: TBase) => class extends superclass {
@@ -45,46 +35,6 @@ const BoardsHandler = <TBase extends Constructor<ServerDataOperatorBase>>(superc
             tableName: BOARD_VIEW,
             prepareRecordsOnly,
         }, 'handleBoardViews');
-    };
-
-    /**
-     * handlePropertyFields: Handler responsible for the Create/Update operations on the PropertyField table.
-     */
-    handlePropertyFields = async ({fields, prepareRecordsOnly = true}: HandlePropertyFieldsArgs): Promise<Model[]> => {
-        if (!fields?.length) {
-            logWarning('An empty or undefined "fields" array has been passed to the handlePropertyFields method');
-            return [];
-        }
-
-        const createOrUpdateRawValues = getUniqueRawsBy({raws: fields, key: 'id'});
-
-        return this.handleRecords({
-            fieldName: 'id',
-            transformer: transformPropertyFieldRecord,
-            createOrUpdateRawValues,
-            tableName: PROPERTY_FIELD,
-            prepareRecordsOnly,
-        }, 'handlePropertyFields');
-    };
-
-    /**
-     * handlePropertyValues: Handler responsible for the Create/Update operations on the PropertyValue table.
-     */
-    handlePropertyValues = async ({values, prepareRecordsOnly = true}: HandlePropertyValuesArgs): Promise<Model[]> => {
-        if (!values?.length) {
-            logWarning('An empty or undefined "values" array has been passed to the handlePropertyValues method');
-            return [];
-        }
-
-        const createOrUpdateRawValues = getUniqueRawsBy({raws: values, key: 'id'});
-
-        return this.handleRecords({
-            fieldName: 'id',
-            transformer: transformPropertyValueRecord,
-            createOrUpdateRawValues,
-            tableName: PROPERTY_VALUE,
-            prepareRecordsOnly,
-        }, 'handlePropertyValues');
     };
 };
